Begin implementation of ML matcher endpoint
